+++ app/Http/Controllers/UserController.php
@@ -6,7 +6,6 @@ use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Arr;
-use Illuminate\Support\Facades\DB;
 use Spatie\Permission\Models\Permission;
 use Spatie\Permission\Models\Role;
 
@@ -14,29 +13,25 @@ class UserController extends Controller
 {
     public function index(Request $request)
     {
-        //$this->authorize('viewAny', User::class);
-        $users = User::all();
+        if(auth()->user()->hasRole('admin')){
+            $users = User::all();
+        } else {
+            $users = User::where('id','!=',1)->get();
+        }
         return view('users.index', compact('users'));
     }
 
     public function create()
     {
-        $this->authorize('create', User::class);
-
-        $roles = User::getRoles();
-
-        return view('users.create',compact('roles' ));
+        return view('users.create');
     }
 
     public function store(Request $request)
     {
-        $this->authorize('create', User::class);
-
         $this->validate($request, [
             'name' => 'required',
             'email' => 'required|email|unique:users,email',
             'password' => 'required|same:confirm-password',
-            'role' => 'required'
         ]);
 
         $input = $request->all();
@@ -44,40 +39,26 @@ class UserController extends Controller
         $input['user_id'] = auth()->id();
 
         $user = User::create($input);
-        $user->assignRole($request->input('role'));
 
-        return redirect()->route('users.index')
-                        ->with('success','User created successfully');
+        return redirect()->route('users.index')->with('success','Registro criado.');
     }
-
     public function show(User $user)
     {
-        //$this->authorize('any', User::class);
         $roles = Role::all();
         $permissions = Permission::all();
         return view('users.role',compact('user', 'roles', 'permissions'));
     }
-
     public function edit(User $user)
     {
-        $this->authorize('update', $user);
-
-        $roles = User::getRoles();
-
-        $userRole = $user->roles->pluck('name','name')->all();
-
-        return view('users.edit',compact('user','roles','userRole', ));
+        return view('users.edit',compact('user'));
     }
 
     public function update(Request $request, User $user)
     {
-        $this->authorize('update', $user);
-
         $this->validate($request, [
             'name' => 'required',
             'email' => 'required|email|unique:users,email,'.$user->id,
             'password' => 'same:confirm-password',
-            'role' => 'required'
         ]);
 
         $input = $request->all();
@@ -89,55 +70,79 @@ class UserController extends Controller
 
         $user->update($input);
 
-        DB::table('model_has_roles')->where('model_id',$user->id)->delete();
-        $user->assignRole($request->input('role'));
-
-        //DB::table('client_user')->whereIn('client_id',$request->input('clients'))->delete();
-        //$user->clients()->sync($request->input('clients'));
-
-        return redirect()->route('users.index')
-                        ->with('success','User updated successfully');
+        return redirect()->route('users.index')->with('success','Registro atualizado.');
     }
 
     public function destroy(User $user)
     {
-        //$this->authorize('update', $user);
         if($user->hasRole('admin')){
             return redirect()->route('users.index')->with('warning','Você é um administrador.');
         }
+
+        if($user->id === auth()->id()){
+            return redirect()->route('users.index')->with('warning','Permissão negada.');
+        }
+
         $user->delete();
+
         return redirect()->route('users.index')->with('success','Usuário removido.');
     }
 
     public function assignRole(Request $request, User $user)
     {
-        if($user->hasRole($request->role)){
-            return redirect()->route('users.show', $user->id)->with('warning','Registro já está adicionado.');
-        }
-        $user->assignRole($request->role);
-        return redirect()->route('users.show', $user->id)->with('success','Registro adicionado.');
-    }
+        $this->authorize('users.roles.edit');
 
-    public function removeRole(User $user, Role $role)
-    {
-        if($user->hasRole($role)){
-            $user->removeRole($role);
-            return redirect()->route('users.show', $user->id)->with('success','Registro removido.');
+        $roles = $user->roles;
+        //dd($roles);
+        //remove roles user
+        if($roles){
+            foreach($roles as $role) {
+                if($user->hasRole($role)){
+                    $user->removeRole($role);
+                }
+            }
         }
-        return redirect()->route('users.show', $user->id)->with('warning','Registro não existe.');
+
+        //dd($request->role);
+        // add roles user
+        if($request->role){
+            foreach($request->role as $role) {
+                if(!$user->hasRole($role)){
+                    $user->assignRole($role);
+                }
+            }
+        }
+
+        return redirect()->route('users.show', $user->id)->with('success','Registro adicionado.');
     }
 
     public function givePermission(Request $request, User $user)
     {
-        if($user->hasPermissionTo($request->permission)){
-            return redirect()->route('users.show', $user->id)->with('warning','Registro já está adicionado.');
+        $this->authorize('users.permissions.edit');
+
+        //dd($user->permissions);
+        if($user->permissions){
+            foreach($user->permissions as $permission){
+                if($user->hasPermissionTo($permission)){
+                    $user->revokePermissionTo($permission);
+                }
+            }
         }
-        $user->givePermissionTo($request->permission);
+
+        //dd($request->permissions);
+        if($request->permissions){
+            foreach($request->permissions as $permission){
+                $user->givePermissionTo($permission);
+            }
+        }
+
         return redirect()->route('users.show', $user->id)->with('success','Registro adicionado.');
     }
 
     public function revokePermission(User $user, Permission $permission)
     {
+        $this->authorize('permissions-update');
+
         if($user->hasPermissionTo($permission)){
             $user->revokePermissionTo($permission);
             return redirect()->route('users.show', $user->id)->with('success','Registro removido.');
